+++ test/test_util.py
@@ -1,6 +1,10 @@
 from __future__ import annotations
+import os
+from pathlib import PureWindowsPath
 import pytest
-from gitmatch import chomp, pathway, trim_trailing_spaces
+from gitmatch import chomp, is_complex_path, pathway, trim_trailing_spaces
+
+ON_WINDOWS = os.name == "nt"
 
 
 @pytest.mark.parametrize(
@@ -56,3 +60,39 @@ def test_pathway(path: str, ways: list[str]) -> None:
 def test_chomp(s: str, chomped: str) -> None:
     assert chomp(s) == chomped
     assert chomp(s.encode("us-ascii")) == chomped.encode("us-ascii")
+
+
+@pytest.mark.parametrize(
+    "path,r",
+    [
+        ("foo", False),
+        ("foo/bar", False),
+        ("foo/bar/", False),
+        ("/foo", True),
+        ("/foo/bar", True),
+        ("/foo/bar/", True),
+    ],
+)
+def test_is_complex_path(path: str, r: bool) -> None:
+    assert is_complex_path(path) == r
+
+
+@pytest.mark.parametrize(
+    "path",
+    [
+        "\\\\?\\C:\\",
+        r"C:\foo\bar",
+        r"\foo\bar",
+        r"C:foo\bar",
+        r"\\?\pictures\kittens",
+        r"\\server\share",
+        r"\\.\BrainInterface",
+        r"\\?\UNC\server\share",
+    ],
+)
+def test_is_complex_path_windows_nonsense(path: str) -> None:
+    if ON_WINDOWS:
+        assert is_complex_path(path)
+    else:
+        assert not is_complex_path(path)
+    assert is_complex_path(PureWindowsPath(path))