+++ CHANGELOG.md
@@ -3,6 +3,8 @@ v0.2.0 (in development)
 - Support Python 3.11 and 3.12
 - Migrated from setuptools to hatch
 - Drop support for Python 3.7
+- `Gitignore.match()` will now raise `InvalidPathError` for Windows paths with
+  anchors
 
 v0.1.0 (2022-07-20)
 -------------------